requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
PyJWT==2.9.0
bcrypt==4.2.0
pytest==8.3.3
